requirements.txt
Pillow==8.4.0
tk==0.1.0
